feat translate added, mobil setting done
