+++ .eslintrc.json
@@ -12,14 +12,38 @@
     "import/resolver": {
       "alias": {
         "map": [
-          ["~components", "./src/components"],
-          ["~config", "./src/config"],
-          ["~data", "./src/data"],
-          ["~hooks", "./src/hooks"],
-          ["~lib", "./src/lib"],
-          ["~pages", "./src/pages"],
-          ["~theme", "./src/theme"],
-          ["~utils", "./src/utils"]
+          [
+            "~components",
+            "./src/components"
+          ],
+          [
+            "~config",
+            "./src/config"
+          ],
+          [
+            "~data",
+            "./src/data"
+          ],
+          [
+            "~hooks",
+            "./src/hooks"
+          ],
+          [
+            "~lib",
+            "./src/lib"
+          ],
+          [
+            "~pages",
+            "./src/pages"
+          ],
+          [
+            "~theme",
+            "./src/theme"
+          ],
+          [
+            "~utils",
+            "./src/utils"
+          ]
         ],
         "extensions": [
           ".js",
@@ -36,8 +60,8 @@
     "prettier/prettier": [
       "error",
       {
-        "endOfLine": "auto",
+        "endOfLine": "auto"
       }
     ]
-  },
+  }
 }
