--- .vscode/extensions.json
@@ -1,3 +1,0 @@
-{
-  "recommendations": ["dbaeumer.vscode-eslint", "eamodio.gitlens", "esbenp.prettier-vscode", "Gruntfuggly.todo-tree"]
-}
