+++ src/components/layout/layout.js
@@ -4,18 +4,12 @@ import React from 'react'
 
 import { Footer } from './footer/footer'
 import { Header } from './header/header'
-import { Profile } from './profile'
 
 export const Layout = ({ children, seo, isLoading = false }) => {
-
   return (
     <>
       {seo && <NextSeo {...seo} />}
       <Flex flexDir='column' minHeight='100vh'>
-        {
-          // <Profile />
-        }
-
         <Header />
         {isLoading ? (
           <Center minH={{ base: 'calc(100vh - 64px)', lg: 'calc(100vh - 100px)' }}>
