+++ src/components/login/oauth-button-group.js
@@ -1,0 +1,28 @@
+import { Button, ButtonGroup, VisuallyHidden } from '@chakra-ui/react'
+
+import { GitHubIcon, GoogleIcon, TwitterIcon } from './provider-icons'
+const providers = [
+  {
+    name: 'Google',
+    icon: <GoogleIcon boxSize='5' />,
+  },
+  {
+    name: 'Twitter',
+    icon: <TwitterIcon boxSize='5' />,
+  },
+  {
+    name: 'GitHub',
+    icon: <GitHubIcon boxSize='5' />,
+  },
+]
+
+export const OAuthButtonGroup = () => (
+  <ButtonGroup variant='outline' spacing='4' width='full'>
+    {providers.map(({ name, icon }) => (
+      <Button key={name} isFullWidth>
+        <VisuallyHidden>Sign in with {name}</VisuallyHidden>
+        {icon}
+      </Button>
+    ))}
+  </ButtonGroup>
+)
