+++ src/hooks/useUser.js
@@ -12,7 +12,6 @@ export const useUser = (redirectTo = '', redirectIfFound = false) => {
         // if no redirect needed, just return (example: already on /dashboard)
         // if user data not yet there (fetch in progress, logged in or not) then don't do anything yet
         if (!redirectTo || !user?.isLoggedIn) return
-
         if (
             // If redirectTo is set, redirect if the user was not found.
             (redirectTo && !redirectIfFound && !user?.isLoggedIn) ||
