--- src/utils/services/index.js
@@ -1,2 +1,0 @@
-export { setToken, getToken } from './token'
-export { getUserFromStorage, removeDataFromLocalStorage } from './user'
