--- src/utils/services/token/index.js
@@ -1,11 +1,0 @@
-const baseUrl = '/api/users'
-
-let token = null
-
-export const setToken = newToken => {
-    token = `Bearer ${newToken}`
-}
-
-export const getToken = () => {
-    return token
-}
