--- src/utils/services/user/index.js
@@ -1,8 +1,0 @@
-export const getUserFromStorage = () => {
-    const user = JSON.parse(localStorage.getItem("loggedUser"))
-    return user
-}
-
-export const removeDataFromLocalStorage = () => {
-    localStorage.removeItem("loggedUser")
-}
